requirements.txt
numpy==1.13.3
tqdm==4.36.1
matplotlib==2.1.1
torchvision==0.4.1
torch==1.2.0
TripletTorch==0.1.2
Pillow==6.2.1
scikit_learn==0.22
